Create READEME.md
Summary of 3 data sets that I am interested in: 
1.	Population Projections: Population projections from 2010 – 2040 (increments of 5 years), broken down by age groups, gender, school age population, and and 65 and older population. Total NYC figures are given, as well projections for each borough. 
2.	Demographics Data Set: I can’t quite view this one as it is a geojson file, but based on the short description that it is ”highly granular demographic data set,” I’m hoping I will find the details of age, income, ethnicity, housing ownership, land use, property value, crime, race, ethnicity, education, disability, housing condition, twitter, and foot traffic of each of each of the boroughs.  
3.	Tree Consensus: Based on what I can view on the excel sheets, this data set is comprised of two sets: one from year 1995 and one from year 2015. Each set indicates the approximate number and location of trees in each borough down to the street address and latitude and longitude lines. 
  
## Data Review 

Selected Data Sources: 
1.	Population Projections 
2.	Demographics data set
3.	Tree Consensus 
** Also interested in: 
4.	Green Infrastructure 
5.	Property valuation and Assessment data 

Supplemental Sources: NYC Open Data, data.gov, healthdata.gov, facebook Graph (to see how people from areas of nyc are sharing with the world), Google Trends (http://www.google.com/trends/explore- Statistics on search volume (as a proportion of total search) for any given term, since 2004.), Topsy (topsy.com - Topsy provides a searchable database of public tweets going back to 2006 as well as several tools to analyze the conversations.)


Questions: 
For the Tree Consensus, are there projections of how many more trees there will be up until 2040 (to match the population projections? 
How can this data be used or displayed to help assess the overall health of trees?

--
As the school age population grows, how can this data be used to meaningfully assist with planning the growth of schools in the city?

--
For the demographic data set: I can’t quite view this yet, but are there any kinds of projections of change in demographics as the population swells and how are they indicated? 
As property values rise, how does that change the demographics of the neighborhoods? What does that look like?